Native references cleanup
