+++ Universal Tomb Launcher/Configuration.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using System.Xml.Serialization;
 using UniversalTombLauncher.Enums;
+using UniversalTombLauncher.Native;
 
 namespace UniversalTombLauncher
 {
