+++ Universal Tomb Launcher/Helpers/InputHelper.cs
@@ -2,6 +2,7 @@
 using System.Drawing;
 using System.Runtime.InteropServices;
 using System.Windows.Forms;
+using UniversalTombLauncher.Native;
 
 namespace UniversalTombLauncher.Helpers
 {
