+++ Universal Tomb Launcher/Native/NativeClasses.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Runtime.InteropServices;
+
+namespace UniversalTombLauncher.Native
+{
+	[ComImport]
+	[Guid("00021401-0000-0000-C000-000000000046")]
+	internal class ShellLink
+	{ }
+}
