+++ Universal Tomb Launcher/Utils/WindowUtils.cs
@@ -2,6 +2,7 @@
 using System.Drawing;
 using System.Runtime.InteropServices;
 using System.Windows.Forms;
+using UniversalTombLauncher.Native;
 
 namespace UniversalTombLauncher.Utils
 {
